{{cookiecutter.repo_name}}/requirements/local.txt
-r base.txt

django-debug-toolbar
pyflakes
django-nose

# LFS
# http://pypi.iqpp.de/django-paypal-0.1.2.lfs-2.tar.gz
-e git+git@github.com:diefenbach/django-lfs.git@feature/django-1.10#egg=lfs
-e git+git@github.com:diefenbach/lfs-contact.git@develop#egg=lfs_contact
-e git+git@github.com:diefenbach/lfs-order-numbers.git@develop#egg=lfs_order_numbers
-e git+git@github.com:diefenbach/lfs-paypal.git@develop#egg=lfs_paypal
-e git+git@github.com:diefenbach/lfs-theme.git@develop#egg=lfs_theme
-e git+git@github.com:diefenbach/django-portlets.git@develop#egg=portlets
-e git+git@github.com:diefenbach/django-reviews.git@develop#egg=reviews
